move the lockout code to Layout.ts, to prevent topbar navigation on lockout
